chore: 소규모 기술부채 일괄 정리 (#137)
* chore: pre-commit 훅을 staged 파일 범위로 제한

- 기존 `git add -u <dir>`가 분할 커밋 시 unstaged 변경을 끌어들이는 문제 해결
- `git diff --cached --diff-filter=ACMR`로 staged 파일만 재스테이지
- `while IFS= read -r` 패턴으로 공백/한글 경로 안전 처리

* chore: .env.prod.sample에 APP_OPENAPI_DOCS_MODE 추가

* refactor: DiscordWebhookClient @RequiredArgsConstructor 적용

* refactor: REQUIRES_NEW TransactionTemplate 공용 Bean으로 분리

- RefreshTokenService를 @RequiredArgsConstructor로 전환

* refactor: DiscordWebhookConfig timeout 캐스팅을 Math.toIntExact로 교체

- Duration overflow 시 ArithmeticException으로 기동 단계에서 조기 감지

* docs: 커밋 본문 불릿 형식 규칙 명시

* fix: TransactionTemplate Bean 오염으로 전 서비스에 REQUIRES_NEW 주입되는 문제 해결

- 기본 transactionTemplate Bean을 TransactionConfig에 추가하여 Spring Boot TransactionAutoConfiguration이 @ConditionalOnMissingBean으로 차단했던 기본 Bean 복구
- 필드명 매칭으로 5개 서비스(CsvUploadService, AdminSystemService, AdminSentenceService, InitialDataSeeder, DailySentenceScheduler)는 REQUIRED 전파, RefreshTokenService는 REQUIRES_NEW 전파로 구분 주입
- 두 Bean 공존 및 각 전파 설정 유지 회귀 테스트 추가

+++ backend/src/main/java/com/gakkaweo/backend/infra/notification/config/DiscordWebhookConfig.java
@@ -12,7 +12,7 @@ public class DiscordWebhookConfig {
 
   @Bean
   RestClient discordWebhookRestClient(DiscordWebhookProperties properties) {
-    int timeoutMillis = (int) properties.getTimeout().toMillis();
+    int timeoutMillis = Math.toIntExact(properties.getTimeout().toMillis());
 
     SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
     factory.setConnectTimeout(timeoutMillis);
